projects/project3-colorful-canvas/src/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.1.78
scipy==1.11.3
matplotlib==3.7.2